+++ src/main/java/com/lxp/LxpApplication.java
@@ -1,7 +1,10 @@
 package com.lxp;
 
+import com.lxp.config.AppConfig;
+
 public class LxpApplication {
 
 	public static void main(String[] args) {
+		new AppConfig().mainView().run();
 	}
 }