+++ src/main/java/com/lxp/controller/CourseController.java
@@ -1,0 +1,13 @@
+package com.lxp.controller;
+
+public class CourseController {
+
+	public void register() { // not implemented
+	}
+
+	public void findAll() { // not implemented
+	}
+
+	public void findById() { // not implemented
+	}
+}